pyproject.toml
[project]
name = "agenticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.1",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.4.8",
    "python-dotenv>=1.1.1",
]
